Updated Map Voting
Added a skip and exit button to the voting view. The exit button
is redundant (as the user can just dismiss the ephemeral message),
but allows the bot a chance to inform the user that their votes
have been saved.

The skip button functions as expected, allowing the user to
skip voting on the current map and move on to the next one.