Custom Emojis
Discord allows the bot to have up to 2,000 custom emojis.
I've added a feature that allows a user to use these custom emojis
in their messages. The bot will delete their message and create a new
one within an embed that contains the custom emoji. If the message is
only the emoji, it will be set as the image of the embed.

To invoke this feature, the user must use the following syntax:
;emoji_name;