Migrate Majority of JSON to ASQLite
Migrated a large amount of the local JSON storage to
asqlite databases for more efficient storage and management.

Updated many functions that relied on the old JSON storage to
utilize the new standards (data types, naming, etc) that are
retrieved from the asqlite databases.

Also, ensured data was properly sorted corrrectly depending
on the context/data.